feat: Add Tags component
